+++ Utilities/ItemHelper.cs
@@ -10,6 +10,71 @@ namespace GeneralImprovements.Utilities
     {
         private static IReadOnlyList<NetworkPrefab> AllPrefabs => NetworkManager.Singleton.NetworkConfig.Prefabs.Prefabs;
 
+        public static GameObject MedStation;
+        public static int MaxHealth;
+
+        public static void CreateMedStation()
+        {
+            if (Plugin.AddHealthRechargeStation.Value && AssetBundleHelper.MedStationPrefab != null)
+            {
+                MedStationItem medStationItem = null;
+
+                if (StartOfRound.Instance.IsHost)
+                {
+                    // If we are the host, spawn it as a network object
+                    Plugin.MLS.LogInfo("Adding medical station to ship.");
+                    MedStation = Object.Instantiate(AssetBundleHelper.MedStationPrefab, new Vector3(2.75f, 3.4f, -16.561f), Quaternion.Euler(-90, 0, 0), StartOfRound.Instance.elevatorTransform);
+                    medStationItem = MedStation.GetComponent<MedStationItem>();
+                    medStationItem.NetworkObject.Spawn();
+                }
+                else
+                {
+                    // If we are a client, try to find it in the scene first
+                    Plugin.MLS.LogInfo("Finding medical station in ship");
+                    medStationItem = Object.FindObjectOfType<MedStationItem>();
+                    MedStation = medStationItem?.gameObject;
+
+                    if (MedStation == null)
+                    {
+                        Plugin.MLS.LogWarning($"Could not find existing med station! Ensure the host has {nameof(Plugin.AddHealthRechargeStation)} set to true.");
+                        return;
+                    }
+                }
+
+                var chargeStation = Object.FindObjectOfType<ItemCharger>().GetComponent<InteractTrigger>();
+                MedStation.GetComponent<AudioSource>().outputAudioMixerGroup = chargeStation.GetComponent<AudioSource>().outputAudioMixerGroup;
+
+                // Add interaction trigger
+                var chargeTriggerCollider = chargeStation.GetComponent<BoxCollider>();
+                chargeTriggerCollider.center = Vector3.zero;
+                chargeTriggerCollider.size = new Vector3(1, 0.8f, 0.8f);
+                var medTrigger = MedStation.transform.Find("Trigger");
+                medTrigger.tag = chargeStation.tag;
+                medTrigger.gameObject.layer = chargeStation.gameObject.layer;
+                var interactScript = medTrigger.gameObject.AddComponent<InteractTrigger>();
+                interactScript.hoverTip = "Heal";
+                interactScript.disabledHoverTip = "(Health Full)";
+                interactScript.hoverIcon = chargeStation.hoverIcon;
+                interactScript.specialCharacterAnimation = true;
+                interactScript.animationString = chargeStation.animationString;
+                interactScript.lockPlayerPosition = true;
+                interactScript.playerPositionNode = chargeStation.playerPositionNode;
+                interactScript.onInteract = new InteractEvent();
+                interactScript.onCancelAnimation = new InteractEvent();
+                interactScript.onInteractEarly = new InteractEvent();
+                interactScript.onInteractEarly.AddListener(_ => medStationItem.HealLocalPlayer());
+
+                // Add scan node
+                var scanNode = MedStation.transform.Find("ScanNode").gameObject.AddComponent<ScanNodeProperties>();
+                scanNode.gameObject.layer = LayerMask.NameToLayer("ScanNode");
+                scanNode.minRange = 0;
+                scanNode.maxRange = 6;
+                scanNode.nodeType = 0;
+                scanNode.headerText = "Med Station";
+                scanNode.subText = "Fully heal yourself";
+            }
+        }
+
         public static void AlterFancyLampPrefab()
         {
             // Add fancy lamp toggle to existing fancy lamp network prefab